KR4K3N-Weapon/docker/requirements.txt
attrs==23.1.0
blinker==1.7.0
click==8.1.7
Flask==3.0.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
MarkupSafe==2.1.3
pycparser==2.21
typing_extensions==4.8.0
Werkzeug==3.0.1
zipp==3.17.0
pyzmq
requests
